Move useTimerControls to its own folder
