+++ src/hooks/useTimerControls.js
